% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_probes_defunct.R
\name{compare_probes_defunct}
\alias{compare_probes_defunct}
\title{THIS FUNCTION IS NOW DEFUNCT}
\usage{
compare_probes_defunct(data_dat, progress = NULL)
}
\arguments{
\item{data_dat}{a text file with the end *score.dat that has the genotypic dosage call output from fitPoly}

\item{progress}{if \code{TRUE} progress shown in console; if \code{FALSE}, no output produced default is TRUE}
}
\value{
Two CSV files. One has the consensus dosage call, other has description of the call. S = same, D = different, O = one, NA = not called
}
\description{
THIS FUNCTION IS NOW DEFUNCT
}
\author{
Jeekin Lau, \email{jeekinlau@gmail.com}
}
