% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosage_to_ACGT_function.R
\name{dosage_to_ACGT}
\alias{dosage_to_ACGT}
\title{Description:  turns dosage based data 0,1,2,3,4 to ACGT with dosage as number of alternative or B_allele
              i.e. if SNP1 A_allele is A and B_allele is T then 0 -> AAAA and 2 -> AATT}
\usage{
dosage_to_ACGT(dosage.file, progress = NULL, separator = NULL)
}
\arguments{
\item{dosage.file}{An output file from the result of compare_probes() or a dosage matrix which rows are markers and columns are individuals}

\item{progress}{if \code{TRUE} progress shown in console; if \code{FALSE}, no output produced default is TRUE}

\item{separator}{what separator used between the letters}
}
\value{
One csv with dosage turned into lettering format
}
\description{
Author: Jeekin Lau
Date last edited: 4/1/2021
}
\details{
input file is dosage file from fitPoly
an assumption of this is that all are same ploidy
only works for 4x right now (may softcode later for other ploidy)

How to use:
    Need a dosage formatted like below:

marker	ind1	ind2	ind3	ind4	ind5
marker1 	4	  3	    4	    4	    3
marker2 	4	  3	    4	    3	    4
marker3 	4	  3	    NA	  NA	  4
marker4 	4	  4	    NA	  4	    3
marker5 	4	  3	    4	    4	    3




NEEDS INTERNET connection to fetch annotation file which contains marker names and A_allele and B_allele for WagRhSNP68k it is "WagRhSNP_annotation.r1.csv"
}
\author{
Jeekin Lau, \email{jeekinlau@gmail.com}
}
