% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitpoly_to_mappoly_input.R
\name{fitPoly_to_MAPpoly_input}
\alias{fitPoly_to_MAPpoly_input}
\title{Prepare data for MAPpoly
Function takes the output from compare_probes() and prepares it for input to use in MAPpoly
Function reads in annotation files and matches the genome position of Rosa chinensis genome (Hibrand-Saint Oyant et al., 2018)
NOTE: must be connected to internet to read in annotation files}
\usage{
fitPoly_to_MAPpoly_input(input_file, p1, p2)
}
\arguments{
\item{input_file}{is a .csv file that has the dosages and Axiom marker names. Most likely an output file of compare_probes() function}

\item{p1}{define parent 1 name}

\item{p2}{define parent 2 name}
}
\value{
a .csv file that has the added columns of genome position and LG of markers. Will contain markers with no known location at bottom of file.
}
\description{
Prepare data for MAPpoly
Function takes the output from compare_probes() and prepares it for input to use in MAPpoly
Function reads in annotation files and matches the genome position of Rosa chinensis genome (Hibrand-Saint Oyant et al., 2018)
NOTE: must be connected to internet to read in annotation files
}
