% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyorigin_imputation.R
\name{polyorigin_imputation}
\alias{polyorigin_imputation}
\title{this function will take the polyorigin postdoseprob.csv output which contains conditional probabilities and outputs}
\usage{
polyorigin_imputation(input.file, number.parents, ploidy, probability)
}
\arguments{
\item{input.file}{name of the file you want to convert from polyorigin}

\item{number.parents}{number or parental genotypes in the file}

\item{ploidy}{number as an integer}

\item{probability}{probability you want to use as a cutoff. recommend to use 0.9}
}
\description{
this function will take the polyorigin postdoseprob.csv output which contains conditional probabilities and outputs
}
